# server/api-doc.json
{"openapi":"3.0.0","paths":{"/api/users/me":{"get":{"operationId":"UsersController_me","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReturnUserDto"}}}}},"tags":["users"]}},"/api/users/logout":{"get":{"operationId":"UsersController_logout","parameters":[],"responses":{"200":{"description":""}},"tags":["users"]}},"/api/users/signup":{"post":{"operationId":"UsersController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReturnUserDto"}}}}},"tags":["users"]}},"/api/users/{id}":{"put":{"operationId":"UsersController_updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReturnUserDto"}}}}},"tags":["users"]},"delete":{"operationId":"UsersController_deleteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReturnUserDto"}}}}},"tags":["users"]}},"/api/users/login":{"post":{"operationId":"UsersController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDto"}}}},"responses":{"200":{"description":""}},"tags":["users"]}},"/api/postings":{"get":{"operationId":"PostingsController_getAllPostings","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Posting"}}}}}},"tags":["postings"]},"post":{"operationId":"PostingsController_createPosting","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/CreatePostingDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Posting"}}}}},"tags":["postings"]}},"/api/postings/category/{category}":{"get":{"operationId":"PostingsController_getPostingsCategory","parameters":[{"name":"category","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["postings"]}},"/api/postings/{id}":{"get":{"operationId":"PostingsController_getPosting","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["postings"]},"put":{"operationId":"PostingsController_editPosting","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePostingDto"}}}},"responses":{"200":{"description":""}},"tags":["postings"]},"delete":{"operationId":"PostingsController_deletePosting","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["postings"]}},"/api/postings/{id}/comments":{"post":{"operationId":"PostingsController_createComment","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCommentDto"}}}},"responses":{"201":{"description":""}},"tags":["postings"]}},"/api/postings/{id}/comments/{commentId}":{"delete":{"operationId":"PostingsController_deleteComment","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}},{"name":"commentId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["postings"]}},"/api/postings/{id}/comments/{commentId}/like":{"get":{"operationId":"PostingsController_likeComment","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}},{"name":"commentId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["postings"]}}},"info":{"title":"Alt Text API","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"ReturnUserDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"email":{"type":"string"},"points":{"type":"number"}},"required":["id","name","email","points"]},"CreateUserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"confirm":{"type":"string"}},"required":["name","email","password","confirm"]},"UpdateUserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"confirm":{"type":"string"}}},"LoginUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"Posting":{"type":"object","properties":{"id":{"type":"number"},"userId":{"type":"number"},"text":{"type":"string"},"imgPath":{"type":"string"},"category":{"type":"string"},"reports":{"type":"number"},"comments":{"type":"array","items":{"$ref":"#/components/schemas/Comment"}},"createdAt":{"format":"date-time","type":"string"}},"required":["id","userId","text","imgPath","category","reports","comments","createdAt"]},"Comment":{"type":"object","properties":{"id":{"type":"number"},"userId":{"type":"number"},"content":{"type":"string"},"likes":{"type":"number"},"likedUsers":{"type":"array","items":{"type":"string"}},"reports":{"type":"number"},"posting":{"$ref":"#/components/schemas/Posting"},"createdAt":{"format":"date-time","type":"string"}},"required":["id","userId","content","likes","likedUsers","reports","posting","createdAt"]},"CreatePostingDto":{"type":"object","properties":{"text":{"type":"string"},"category":{"type":"string"},"image":{"type":"object"}},"required":["text","category","image"]},"UpdatePostingDto":{"type":"object","properties":{"text":{"type":"string"},"category":{"type":"string"},"image":{"type":"object"}}},"CreateCommentDto":{"type":"object","properties":{"content":{"type":"string"}},"required":["content"]}}}}